기본 프로필일 경우 프로필 삭제하지 않음

+++ src/main/java/com/example/let/service/impl/UserServiceImpl.java
@@ -144,7 +144,9 @@ public class UserServiceImpl implements UserService {
         entryMapper.deleteByUserId(user.getIdx());
         accessMapper.delete(id);
         cardMapper.deleteByUserId(user.getIdx());
-        fileMapper.deleteFile(user.getImage());
+        if (user.getImage() != 0) {
+            fileMapper.deleteFile(user.getImage());
+        }
         userMapper.delete(id);
     }
